refactor: Jaeger tracing 설정 업데이트 및 파일 이동
