+++ src/tracing.ts
@@ -1,7 +1,6 @@
 import { NodeSDK } from '@opentelemetry/sdk-node';
-import { getNodeAutoInstrumentations } from '@opentelemetry/auto-instrumentations-node';
 import { OTLPTraceExporter } from '@opentelemetry/exporter-trace-otlp-http';
-import { PinoInstrumentation } from '@opentelemetry/instrumentation-pino';
+import { getNodeAutoInstrumentations } from '@opentelemetry/auto-instrumentations-node';
 
 if (!process.env.JAEGER_ENDPOINT) {
   throw new Error('JAEGER_ENDPOINT environment variable is not set');
@@ -14,7 +13,26 @@ const sdk = new NodeSDK({
     headers: {},
   }),
   serviceName: 'main-game-service',
-  instrumentations: [getNodeAutoInstrumentations(), new PinoInstrumentation()],
+  instrumentations: [
+    getNodeAutoInstrumentations({
+      '@opentelemetry/instrumentation-kafkajs': {
+        enabled: true,
+        producerHook: (span, { topic, message }) =>
+          span.setAttributes({
+            'kafka.topic': topic,
+            'kafka.message.key': message.key ? message.key.toString() : undefined,
+            'kafka.message.value': message.value ? message.value.toString() : undefined,
+          }),
+        consumerHook: (span, { topic, message }) =>
+          span.setAttributes({
+            'kafka.topic': topic,
+            'kafka.message.key': message.key ? message.key.toString() : undefined,
+            'kafka.message.value': message.value ? message.value.toString() : undefined,
+          }),
+      },
+    }),
+  ],
 });
 
 sdk.start();
+console.log('[DEBUG] OpenTelemetry SDK started with Jaeger exporter');
